+++ examples/shallow_encode_decode/shallow_encode_decode.c
@@ -16,8 +16,10 @@
 /// Example showing how to encode and decode symbols that are stored in
 /// non-contiguous memory buffers.
 
-void trace_callback(const char* zone, const char* data)
+void trace_callback(const char* zone, const char* data, void* context)
 {
+    (void) context;
+
     if (strcmp(zone, "decoder_state") == 0)
     {
         printf("%s:\n", zone);
@@ -36,7 +38,7 @@ int main()
     uint32_t symbol_size = 100;
 
     // Here we select the coding code_type we wish to use
-    int32_t code_type = kodo_full_rlnc;
+    int32_t code_type = kodo_full_vector;
 
     // Here we select the finite field to use common choices are
     // kodo_binary, kodo_binary8, kodo_binary16
@@ -120,9 +122,9 @@ int main()
     // }
 
     // Install a custom trace function for the decoder (if tracing is enabled)
-    if (kodo_has_trace(decoder))
+    if (kodo_has_set_trace_callback(decoder))
     {
-        kodo_trace_callback(decoder, trace_callback);
+        kodo_set_trace_callback(decoder, trace_callback, NULL);
     }
 
     while (!kodo_is_complete(decoder))