+++ src/kodoc/kodo_coder.hpp
@@ -19,16 +19,21 @@ extern "C"
         virtual uint32_t rank() const = 0;
         virtual uint32_t symbol_size() const = 0;
         virtual uint32_t symbols() const = 0;
-        virtual bool is_symbol_pivot(uint32_t) const = 0;
+        virtual bool is_symbol_pivot(uint32_t index) const = 0;
 
-        virtual void set_symbols(
-            uint8_t* data, uint32_t size) = 0;
+        virtual void set_symbols(uint8_t* data, uint32_t size) = 0;
 
-        virtual void set_symbol(
-            uint32_t index, uint8_t* data, uint32_t size) = 0;
+        virtual void set_symbol(uint32_t index, uint8_t* data,
+            uint32_t size) = 0;
 
-        virtual bool has_trace() const = 0;
-        virtual void trace(kodo_trace_callback_t) = 0;
+        virtual bool has_set_trace_stdout() const = 0;
+        virtual bool has_set_trace_callback() const = 0;
+        virtual bool has_set_trace_off() const = 0;
+
+        virtual void set_trace_stdout() = 0;
+        virtual void set_trace_callback(kodo_trace_callback_t callback,
+            void* context) = 0;
+        virtual void set_trace_off() = 0;
 
         virtual bool has_feedback_size() const = 0;
         virtual uint32_t feedback_size() const = 0;