+++ src/kodoc/kodo_encoder.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "coder.hpp"
+#include "kodo_coder.hpp"
 
 extern "C"
 {
@@ -22,9 +22,9 @@ extern "C"
         virtual void set_systematic_off() = 0;
         virtual void read_feedback(uint8_t* feedback) = 0;
 
-        virtual uint32_t write_symbol(
-            uint8_t* symbol_data, uint8_t* coefficients) = 0;
-        virtual uint32_t write_uncoded_symbol(
-            uint8_t* symbol_data, uint32_t index) = 0;
+        virtual uint32_t write_symbol(uint8_t* symbol_data,
+            uint8_t* coefficients) = 0;
+        virtual uint32_t write_uncoded_symbol(uint8_t* symbol_data,
+            uint32_t index) = 0;
     };
 }