+++ src/kodoc/kodo_new_decoder_factory.cpp
@@ -9,10 +9,10 @@
 #include <cstdint>
 #include <cassert>
 
-#include <kodo/rlnc/full_vector_codes.hpp>
-#include <kodo/rlnc/on_the_fly_codes.hpp>
-#include <kodo/rlnc/seed_codes.hpp>
+#include <kodo/rlnc/full_vector_decoder.hpp>
+#include <kodo/rlnc/on_the_fly_decoder.hpp>
 #include <kodo/rlnc/sliding_window_decoder.hpp>
+#include <kodo/rlnc/perpetual_decoder.hpp>
 
 #include "create_factory_wrapper.hpp"
 #include "decoder_factory_wrapper.hpp"
@@ -31,7 +31,7 @@ kodo_new_decoder_factory(int32_t code_type, int32_t finite_field,
 
     kodo_factory_t factory = 0;
 
-    if (code_type == kodo_full_rlnc)
+    if (code_type == kodo_full_vector)
     {
         factory = create_factory_wrapper<
             decoder_factory_wrapper, full_vector_decoder>(
@@ -47,6 +47,12 @@ kodo_new_decoder_factory(int32_t code_type, int32_t finite_field,
     {
         factory = create_factory_wrapper<
             decoder_factory_wrapper, sliding_window_decoder>(
+                finite_field, max_symbols, max_symbol_size, trace_mode);
+    }
+    else if (code_type == kodo_perpetual)
+    {
+        factory = create_factory_wrapper<
+            decoder_factory_wrapper, perpetual_decoder>(
                 finite_field, max_symbols, max_symbol_size, trace_mode);
     }