+++ src/kodoc/kodo_new_shallow_encoder_factory.cpp
@@ -31,25 +31,25 @@ kodo_new_shallow_encoder_factory(int32_t code_type, int32_t finite_field,
 
     kodo_factory_t factory = 0;
 
-    if (code_type == kodo_full_rlnc)
+    if (code_type == kodo_full_vector)
     {
         factory = create_factory_wrapper<
             encoder_factory_wrapper, shallow_full_vector_encoder>(
                 finite_field, max_symbols, max_symbol_size, trace_mode);
     }
-    else if (code_type == kodo_sparse_full_rlnc)
+    else if (code_type == kodo_sparse_full_vector)
     {
         factory = create_factory_wrapper<
             sparse_encoder_factory_wrapper, shallow_sparse_full_vector_encoder>(
                 finite_field, max_symbols, max_symbol_size, trace_mode);
     }
-    else if (code_type == kodo_seed_rlnc)
+    else if (code_type == kodo_seed)
     {
         factory = create_factory_wrapper<
             encoder_factory_wrapper, shallow_seed_encoder>(
                 finite_field, max_symbols, max_symbol_size, trace_mode);
     }
-    else if (code_type == kodo_sparse_seed_rlnc)
+    else if (code_type == kodo_sparse_seed)
     {
         factory = create_factory_wrapper<
             sparse_encoder_factory_wrapper, shallow_sparse_seed_encoder>(