+++ test/src/test_helper.hpp
@@ -49,14 +49,16 @@ inline void test_combinations(
 
     std::vector<int32_t> code_types =
     {
-        kodo_full_rlnc,
+        kodo_full_vector,
         kodo_on_the_fly,
-        kodo_sliding_window
+        kodo_sliding_window,
+        kodo_perpetual
     };
 
     std::vector<int32_t> fields =
     {
         kodo_binary,
+        kodo_binary4,
         kodo_binary8,
         kodo_binary16
     };